Splitting up code for different versions of martha to hopefully help improve testability
